Avoid reconcile tight loop
A resource will be queued whenever it changes, so if you change each
resource every time you see it, you will spin through all of them
continually.

To avoid that situation, this commit changes the reconcile logic so
that the resource status is updated only

 - if it actually make a commit; or,
 - if `interval` has passed since it last updated the status

In a steady state, this means it'll run twice (once for whatever
triggered it, and once because it updated itself) then be requeued
`interval` later.