Ensure LastAutomationRun only records successes
The `LastAutomationRun` field records the last time the automation
completed, whether it made changes or not. The purpose is to inform
users, so they can check against their expectations.

It should only be set once the automation run has completed without
error, otherwise it will provide false reassurance that things are
happening when they are not.

Signed-off-by: Michael Bridgen <michael@weave.works>