Line up $IMG in Makefile with convention
This changes the value of $IMG in the Makefile to line up with other
GitOps toolkit controllers. The main effect is the the `kustomize edit
image` commands used in the deploy and dev-deploy targets will now set
the image tag, rather than let it stay at whatever the last release
was.

Signed-off-by: Michael Bridgen <michael@weave.works>