Add a note to SourceRef to address Image FAQ
The most common issue for ImageUpdateAutomation questions in #flux and
elsewhere seems to be users who do not realize they cannot put the Image
automation resources into a different namespace than the GitRepository.

Not sure what else we can do to make this information more prominent,
but if users check the API docs, they should definitely be able to find
this information here.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>