Update tests to be compatible with CR > v0.10.0
In controller-runtime v0.10.0, the client is updated to clean any stale
data in the target object when performing any operation. This results in
test failure for the code that constructs an object with both spec and
status, and creates the object and updates status it with the same
object. The fix is to set the status separately on the object before
updating it.

Refer: https://github.com/kubernetes-sigs/controller-runtime/pull/1640

Signed-off-by: Sunny <darkowlzz@protonmail.com>