Add reconciler sharding capability based on label selector
The new `--watch-label-selector` lets operators provide a label to the
controller manager which in turn uses it to reconcile only those resources
`ImageUpdateAutomations` that match the given label.

This way operators can deploy multiple instances of IAC, each reconciling a
distinct set of resources based on their labels and effectively scale IAC
horizontally.

closes #503

Signed-off-by: Max Jonas Werner <mail@makk.es>