Remove dependence on kustomize/setters2
The package setters2 has been removed from kustomize:
https://github.com/kubernetes-sigs/kustomize/pull/5291

This commit removes the need to import setters2, by reproducing the
last link with that code, which is a minor parsing helper.

I have not changed the comment explaining what was changed from the
original, since it's still accurate. The parsing func is a buried
detail.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>